test(publicformauthservice/test): deleted tests in myinfo service and added to publicformauthservice
